+++ Plugin.Maui.Biometric/Plugin.Maui.Biometric/Enums/AuthenticationType.cs
