+++ Plugin.Maui.Biometric/Plugin.Maui.Biometric/Enums/AuthenticatorStrength.cs
