+++ Plugin.Maui.Biometric/Plugin.Maui.Biometric/Enums/BiometricHwStatus.cs
