+++ Plugin.Maui.Biometric/Plugin.Maui.Biometric/Enums/BiometricResponseStatus.cs
