+++ Plugin.Maui.Biometric/Plugin.Maui.Biometric/Enums/BiometricType.cs
