+++ Plugin.Maui.Biometric/Plugin.Maui.Biometric/Enums/BlockMode.cs
