+++ Plugin.Maui.Biometric/Plugin.Maui.Biometric/Enums/CryptoOperation.cs
