+++ Plugin.Maui.Biometric/Plugin.Maui.Biometric/Enums/Digest.cs
