+++ Plugin.Maui.Biometric/Plugin.Maui.Biometric/Enums/KeyAlgorithm.cs
