+++ Plugin.Maui.Biometric/Plugin.Maui.Biometric/Enums/Padding.cs
