+++ Plugin.Maui.Biometric/Plugin.Maui.Biometric/Models/AuthenticationRequest.cs
