+++ Plugin.Maui.Biometric/Plugin.Maui.Biometric/Models/AuthenticationResponse.cs
