+++ Plugin.Maui.Biometric/Plugin.Maui.Biometric/Models/CryptoKeyOptions.cs
