+++ Plugin.Maui.Biometric/Plugin.Maui.Biometric/Models/KeyOperationResult.cs
