+++ Plugin.Maui.Biometric/Plugin.Maui.Biometric/Models/SecureAuthenticationRequest.cs
