+++ Plugin.Maui.Biometric/Plugin.Maui.Biometric/Models/SecureAuthenticationResponse.cs
